restructure html layout
