+++ client/src/components/HomeFeed/HomeFeed.js
@@ -1,16 +1,13 @@
 import React from "react";
 import styled from "styled-components";
-import Tweet from "../Tweet";
+import TweetFeed from "../Tweet";
 
-const HomeWrapper = styled.div`
-  padding-left: 30px;
-  border-left: 1.5px solid #f0f0f5;
-`;
+const HomeWrapper = styled.div``;
 export default function HomeFeed() {
   return (
     <HomeWrapper>
       <h1>Home</h1>
-      <Tweet />
+      <TweetFeed />
     </HomeWrapper>
   );
 }
